Move colors to config/colors
Caleydo/cohort#693
