+++ src/Taskview/tasks/Details.ts
@@ -3,7 +3,7 @@ import { select } from 'd3v7';
 import * as LineUpJS from 'lineupjs';
 import { Cohort } from '../../Cohort';
 import { ICohort } from '../../CohortInterfaces';
-import { colors } from '../../colors';
+import { colors } from '../../config/colors';
 import { IAttribute } from '../../data/Attribute';
 import { getCohortData } from '../../rest';
 import { CohortColorSchema, getAnimatedLoadingText } from '../../util';
